requirements.txt
docopt==0.6.2
ducktoolkit==1.0.3
factordb-pycli==1.3.0
Flask==2.0.3
librosa==0.9.2
matplotlib==3.5.2
numpy==1.21.5
oletools==0.60.1
Pillow==9.2.0
pypcapkit==0.16.2
pwnlib==1.2.1
pwntools==4.8.0
pycryptodome==3.15.0
pydub==0.25.1
pyUnicodeSteganography==0.0.1
r2pipe==1.7.3
scapy==2.4.5
scipy==1.8.1
stegano==0.10.2
stegoveritas==1.9
stegpy==0.0.1
tinyscript==1.26.11
Werkzeug==2.0.2